requirements.txt
# ConsumerSim dependencies
# Install with: pip install -r requirements.txt

# Core data
pandas>=2.0
numpy>=1.24
pyreadstat>=1.2        # read .dta Stata files
openpyxl>=3.1          # read .xlsx Excel files
pyyaml>=6.0
requests>=2.31         # HTTP clients for collectors
pytrends>=4.9          # Google Trends collector

# ML behavior model
scikit-learn>=1.3
xgboost>=2.0           # optional, used if model_type=xgboost

# LLM
openai>=1.0            # OpenAI API (also compatible with vLLM local endpoint)

# Validation
scipy>=1.11
pydantic>=2.0          # response validation

# Visualization
matplotlib>=3.7

# Optional: local LLM inference (GPU required)
# vllm>=0.3.0
